// src/modules/applications/Transactions/models/CryptoOrder.ts
export type CryptoOrderStatus = 'completed' | 'pending' | 'failed';

export type CryptoOrderType = {
  id: string;
  status: CryptoOrderStatus;
  orderDetails: string;
  orderDate: number;
  orderID: string;
  sourceName: string;
  sourceDesc: string;
  amountCrypto: number;
  amount: number;
  cryptoCurrency: string;
  currency: string;
};